config/requirements.txt
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2.4
django-braces==1.13.0
django-filter==2.2.0
django-filters==0.2.1
django-oauth-toolkit==1.2.0
django-photologue==3.9
django-rest-framework-social-oauth2==1.1.0
django-sortedm2m==1.5.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
drf-yasg==1.16.1
ExifRead==2.1.2
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==6.1.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
ruamel.yaml==0.16.1
ruamel.yaml.clib==0.1.2
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
